Switch to f32 throughout
While the performance gains are modest (~10%), which is why I didn't do
this earlier, I underestimated the amount of memory that it takes up,
especially with larger images. I also considered making the code generic
over f32/f64 a while back, but it was a huge pain and also made things
slower, so it's best to just convert everything to f32 and be done with
it.